requirements.txt
fastapi
uvicorn
slack-sdk
python-dotenv
pydantic
requests
anyio
certifi
charset-normalizer
click
fastapi
httpcore
httpx
idna
jsonpatch
jsonpointer
langchain
langchain-core
langchain-text-splitters
langsmith
orjson
packaging
pydantic
pydantic-settings
pydantic_core
python-dateutil
python-dotenv
PyYAML
requests
requests-toolbelt
sfdc-llm-gateway-python-sdk @ file:///Users/vvarunraju/Documents/Projects/slackBotHackathon/sfdc_llm_gateway_python_sdk-1.0.4-py3-none-any.whl#sha256=54693754a30a8d81f5230dcb85ecdfab89de9d012f32ae5888c24bb4144f90ff
six
slack-sdk
sniffio
SQLAlchemy
starlette
tenacity
typing-inspection
typing_extensions
urllib3
uvicorn
zstandard
